requirements.txt
robotframework
robotframework-requests
robotframework-sshlibrary
ping3